requirements.txt
pathspec
python-magic
python-magic-bin; sys_platform == 'win32'